Add metadata for extra emu config including media
An extra per-instance emulator config XML file can be specified to
specify which of the files is a game file, which peripheral it's loaded
into, what human name to call it, etc.

Here, we load it and process its contents, adding the files it requests
to the download queue and appending information to the emulator object
for outside scripts to retrieve.

So far, only implemented for MAME. We also add an initial prototype MAME
Lua file as a placeholder for the eventual script needed to change media
out.